+++ app/src/test/java/com/aowen/monolith/fakes/data/BuildsDto.kt
@@ -1,6 +1,7 @@
 package com.aowen.monolith.fakes.data
 
 import com.aowen.monolith.data.BuildDto
+import com.aowen.monolith.data.GameVersionDto
 import com.aowen.monolith.data.ModuleDto
 
 val fakeModuleDto = ModuleDto(
@@ -11,6 +12,15 @@ val fakeModuleDto = ModuleDto(
     item4Id = 304,
     item5Id = 305,
     item6Id = 306
+)
+
+val fakeGameVersionDto = GameVersionDto(
+    id = 1,
+    name = "1.0.0",
+    release = "2022-01-01T00:00:00Z",
+    displayBadge = true,
+    createdAt = "2022-01-01T00:00:00Z",
+    updatedAt = "2022-01-01T00:00:00Z"
 )
 
 val fakeBuildDto = BuildDto(
@@ -31,6 +41,7 @@ val fakeBuildDto = BuildDto(
     createdAt = "2022-01-01T00:00:00Z",
     updatedAt = "2022-01-01T00:00:00Z",
     author = "Author",
-    modules = listOf(fakeModuleDto)
+    modules = listOf(fakeModuleDto),
+    gameVersion = fakeGameVersionDto
 )